ds-nn-m6/requirements.txt
# FastAPI and server
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.2
typing-extensions>=4.8.0

# Document Processing
python-docx>=0.8.11
PyPDF2==3.0.1
PyMuPDF>=1.23.8
pdfminer.six>=20221105
pdf2docx>=2.1.0

# AWS and Environment
boto3>=1.28.44
python-dotenv>=1.0.0

# Additional Dependencies
requests>=2.31.0
aiohttp>=3.9.1

# URL and Text Processing
urllib3>=2.0.7
regex>=2023.10.3
